Refactor(APP): 컴포넌트 및 훅 구조 개선
- App.tsx를 EventForm, CalendarView, EventList 컴포넌트로 분리하여 단일 책임 원칙을 적용
- 분리된 각 컴포넌트에 대한 단위 테스트를 추가하여 독립적인 렌더링을 검증
- 여러 훅을 useEventManager로 통합하여 App.tsx의 복잡도를 낮추고 상태 관리를 중앙화
- useEventManager 훅에 대한 단위 테스트를 추가
- 과정 기록 문서를 docs/ 디렉토리로 이동하여 정리
